board_inspector: resume a device if it is not in D0
It is seen that the BAR in the PCI configuration space of a device can be
cleared when the device is put to D3. This patch resumes a device not in D0
before parsing its configuration space in order to collect accurate
information.

This patch is added in v2 of the series.

Tracked-On: #6287

+++ misc/config_tools/board_inspector/pcieparser/caps.py
@@ -67,8 +67,8 @@ class PowerManagement(cdata.Struct, Capability):
         ('data_select', ctypes.c_uint16, 4),
         ('data_scale', ctypes.c_uint16, 2),
         ('pme_status', ctypes.c_uint16, 1),
-        ('reserved3', ctypes.c_uint16, 6),
-        ('undefined', ctypes.c_uint16, 2),
+        ('reserved3', ctypes.c_uint8, 6),
+        ('undefined', ctypes.c_uint8, 2),
         ('data', ctypes.c_uint8),
     ]
